fix: working

+++ abc/abc-main/src/main/java/com/example/abc/service/impl/AbcServiceImpl.java
@@ -2,9 +2,22 @@ package com.example.abc.service.impl;
 
 import com.example.abc.domain.dto.AbcDTO;
 import com.example.abc.domain.entity.Abc;
+import com.example.abc.eventuate.AbcTramMessageConfig;
+import com.example.abc.eventuate.event.AbcInsertEvent;
+import com.example.abc.eventuate.event.AbcUpdateEvent;
 import com.example.abc.repository.AbcRepository;
 import com.example.abc.service.AbcService;
 import com.example.abc.service.mapper.AbcMapper;
+import com.example.ecsp.common.util.JsonUtil;
+import com.example.xyz.domain.dto.XyzDTO;
+import com.example.xyz.domain.dto.XyzDetailDTO;
+import com.example.xyz.domain.entity.Xyz;
+import com.example.xyz.eventuate.XyzTramMessageConfig;
+import com.example.xyz.eventuate.command.XyzRegisterCommand;
+import io.eventuate.common.json.mapper.JSonMapper;
+import io.eventuate.tram.commands.producer.CommandProducer;
+import io.eventuate.tram.events.common.DomainEvent;
+import io.eventuate.tram.events.publisher.DomainEventPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -12,7 +25,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collections;
 import java.util.Optional;
+import java.util.Set;
 
 /**
  * Service Implementation for managing {@link com.example.abc.domain.entity.Abc}.
@@ -27,12 +42,36 @@ public class AbcServiceImpl implements AbcService {
 
     private final AbcMapper abcMapper;
 
+    private final DomainEventPublisher domainEventPublisher;
+    private final CommandProducer commandProducer;
+
     @Override
     public AbcDTO save(AbcDTO AbcDTO) {
         log.debug("Request to save Abc : {}", AbcDTO);
         Abc abc = abcMapper.toEntity(AbcDTO);
         abc = abcRepository.save(abc);
         AbcDTO result = abcMapper.toDto(abc);
+
+        log.debug("abc = {}", abc);
+        log.debug("abc.toJson1 = {}", JsonUtil.toJson(abc));
+        log.debug("abc.toJson2 = {}", JSonMapper.toJson(abc));
+
+        // 도메인 이벤트 저장
+        DomainEvent domainEvent = new AbcInsertEvent(abc);
+        domainEventPublisher.publish(Abc.class.getName(), abc.getId(), Collections.singletonList(domainEvent));
+
+        // 등록시 다른 서비스에 Command를 실행해야 하는 경우
+        commandProducer.send(XyzTramMessageConfig.commandChannel,
+                new XyzRegisterCommand(XyzDTO.builder()
+                        .name("abc")
+                        .age(30L)
+                        .status(Xyz.XyzStatus.STANDBY)
+                        .xyzDetails(Set.of(XyzDetailDTO.builder().attrName("data").attrValue(abc.getData()).build()))
+                        .build()),
+                AbcTramMessageConfig.replyChannel,
+                Collections.emptyMap() // header
+        );
+
         return result;
     }
 
@@ -42,6 +81,11 @@ public class AbcServiceImpl implements AbcService {
         Abc abc = abcMapper.toEntity(AbcDTO);
         abc = abcRepository.save(abc);
         AbcDTO result = abcMapper.toDto(abc);
+
+        // 도메인 이벤트 저장
+        DomainEvent domainEvent = new AbcUpdateEvent(abc);
+        domainEventPublisher.publish(Abc.class.getName(), abc.getId(), Collections.singletonList(domainEvent));
+
         return result;
     }
 
@@ -53,6 +97,11 @@ public class AbcServiceImpl implements AbcService {
             .findById(AbcDTO.getId())
             .map(existingAbc -> {
                 abcMapper.partialUpdate(existingAbc, AbcDTO);
+
+                // 도메인 이벤트 저장
+                DomainEvent domainEvent = new AbcUpdateEvent(existingAbc);
+                domainEventPublisher.publish(Abc.class.getName(), existingAbc.getId(), Collections.singletonList(domainEvent));
+
                 return existingAbc;
             })
             .map(abcRepository::save)
@@ -93,10 +142,19 @@ public class AbcServiceImpl implements AbcService {
     public void delete(String id) {
         log.debug("Request to delete Abc : {}", id);
 
+        Abc abc = abcRepository.findById(id).orElse(null);
+
+        if(abc == null) return;
+
 //        Specification<Abc> spec = AbcRepository.Spec.instance().equal("chatRoom", "id", chatRoomId)
 //                .and(AbcRepository.Spec.instance().equal("id", id));
 //        abcRepository.delete(spec);
 
         abcRepository.deleteById(id);
+
+        // 도메인 이벤트 저장
+        DomainEvent domainEvent = new AbcUpdateEvent(abc);
+        domainEventPublisher.publish(Abc.class.getName(), abc.getId(), Collections.singletonList(domainEvent));
+
     }
 }
